// types.ts
export enum TaxStatus {
  RESPONSABLE_INSCRIPTO = 'Responsable Inscripto',
  RESPONSABLE_INSCRIPTO_AGENTE = 'Responsable Inscripto Agente Percepción',
  CONSUMIDOR_FINAL = 'Consumidor final',
  MONOTRIBUTO = 'Monotributo',
  SUJETO_NO_CATEGORIZADO = 'Sujeto NO categorizado',
  EXENTO = 'Exento'
}
// ...
export interface CoopProsumidorGDData {
  taxStatus: TaxStatus | '';
  cuotaServicio: number;
  cargoCapPico: number;
  cargoCapFueraPico: number;
  cargoCapPotAdquirida: number;
  tarifaPico: number;
  tarifaResto: number;
  tarifaValle: number;
  bonificacionFactorPotencia: number;
  reconCoopEnergiaRecibida: number;
  subtotalBasico: number;
  subtotalGeneral: number;
  reconGSFEnergiaGenerada: number;
  totalPagar: number;
  entregadaPico: number;
  entregadaResto: number;
  entregadaValle: number;
  generadaPico: number;
  generadaResto: number;
  generadaValle: number;
  recibidaPico: number;
  recibidaResto: number;
  recibidaValle: number;
}